Refactors PostDetails screen
