+++ Incopo/View/ImageCell.swift
@@ -1,0 +1,13 @@
+//
+//  ImageCell.swift
+//  Incopo
+//
+//  Created by Laurentiu Ile on 25.02.2021.
+//
+
+import UIKit
+
+class ImageCell: UITableViewCell {
+
+    @IBOutlet weak var postImageView: UIImageView!
+}
