+++ Incopo/View/PostDetailCell.swift
@@ -1,0 +1,21 @@
+//
+//  PostDetailCell.swift
+//  Incopo
+//
+//  Created by Laurentiu Ile on 24.02.2021.
+//
+
+import UIKit
+
+class PostDetailCell: UITableViewCell {
+
+    @IBOutlet weak var titleLabel: UILabel!
+    @IBOutlet weak var postContentTextView: UITextView!
+    @IBOutlet weak var byAuthorLabel: UILabel!
+    @IBOutlet weak var likeButton: UIButton!
+    @IBOutlet weak var noLikesLabel: UILabel!
+    @IBOutlet weak var commentsButton: UIButton!
+    @IBOutlet weak var noCommentsLabel: UILabel!
+    @IBOutlet weak var favoriteButton: UIButton!
+    
+}
